requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
fastapi==0.103.1
h11==0.14.0
idna==3.4
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
python-http-client==3.3.7
sendgrid==6.10.0
sniffio==1.3.0
starkbank-ecdsa==2.2.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
